refactor: move permissionTypes to /backend
